requirements.txt
absl-py==2.1.0
accelerate==1.1.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
compel==2.0.3
contourpy==1.3.0
controlnet-aux==0.0.9
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
# diffusers @ git+https://github.com/huggingface/diffusers.git@13e8fdecda91e27e40b15fa8a8f456ade773e6eb
# dwpose==1.0.1
einops==0.8.0
executing==2.1.0
filelock==3.16.1
Flask==3.1.0
Flask-Login==0.6.3
flask-sqlacodegen==2.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
glob2==0.7
greenlet==3.1.1
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
inflect==7.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
#-e git+https://github.com/Modegen/LookBuilderPipeline.git@54b2148ffd2c97e2565919d017a1e56c2f4e8a73#egg=LookBuilderPipeline
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mediapipe==0.10.18
ml_dtypes==0.5.0
more-itertools==10.5.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optimum-quanto==0.2.1
packaging==24.1
parso==0.8.4
peft==0.13.2
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
# sd_embed @ git+https://github.com/xhinker/sd_embed.git@f2a2c135f398055e0cdbe34b80dd1dd08d59acc6
# sentencepiece==0.2.0
setuptools==75.3.0
six==1.16.0
sounddevice==0.5.1
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.3
tifffile==2024.9.20
timm==0.6.7
tokenizers==0.19.1
torch==2.2.2
torchao==0.1
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
transformers==4.44.0
typeguard==4.4.1
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
zipp==3.20.2

pytest

torch
mediapipe
accelerate
bitsandbytes
numpy
sqlalchemy
huggingface-hub
sentencepiece==0.2.0
diffusers @ git+https://github.com/huggingface/diffusers.git
python-dotenv
tokenizers #==0.19
protobuf
compel
psycopg2
dwpose @ git+https://github.com/Modegen/DWPose.git
torchao
sd_embed @ git+https://github.com/xhinker/sd_embed.git
glob2
bitsandbytes
controlnet_aux #==0.0.6
opencv-python
# flash-attn #--no-build-isolation
